+++ backend/canisters/community/api/src/queries/messages_by_message_index.rs
@@ -20,5 +20,7 @@ pub enum Response {
     UserNotInChannel,
     ChannelNotFound,
     ThreadNotFound,
+    UserSuspended,
+    UserLapsed,
     ReplicaNotUpToDateV2(TimestampMillis),
 }