+++ backend/canisters/community/api/src/updates/unfollow_thread.rs
@@ -21,4 +21,5 @@ pub enum Response {
     UserNotInChannel,
     UserSuspended,
     CommunityFrozen,
+    UserLapsed,
 }