+++ backend/canisters/group/impl/src/updates/remove_reaction.rs
@@ -35,6 +35,7 @@ fn remove_reaction_impl(args: Args, state: &mut RuntimeState) -> Response {
             AddRemoveReactionResult::UserNotInGroup => CallerNotInGroup,
             AddRemoveReactionResult::NotAuthorized => NotAuthorized,
             AddRemoveReactionResult::UserSuspended => UserSuspended,
+            AddRemoveReactionResult::UserLapsed => UserLapsed,
         }
     } else {
         CallerNotInGroup