+++ backend/integration_tests/src/client/user.rs
@@ -133,6 +133,7 @@ pub mod happy_path {
                 rules: Rules::default(),
                 events_ttl: None,
                 gate: None,
+                gate_config: None,
             },
         );
 
@@ -163,6 +164,7 @@ pub mod happy_path {
                 permissions: None,
                 rules: Rules::default(),
                 gate: None,
+                gate_config: None,
                 default_channels,
                 default_channel_rules: None,
                 primary_language: "en".to_string(),